Refactor & document supported form fields
